// src/lib/api/models/updateLeadDtoStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */

export type UpdateLeadDtoStatus = typeof UpdateLeadDtoStatus[keyof typeof UpdateLeadDtoStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UpdateLeadDtoStatus = {
  NEW: 'NEW',
  CONTACTED: 'CONTACTED',
  INTERESTED: 'INTERESTED',
  NEGOTIATING: 'NEGOTIATING',
  CONVERTED: 'CONVERTED',
  LOST: 'LOST',
  RECYCLED: 'RECYCLED',
} as const;